+++ disk.h
@@ -1,5 +1,5 @@
 #ifndef DISK_H_
 #define DISK_H_
 
-int disk_read(int id, void *buf, unsigned offset, int count);
+size_t disk_read(int id, void *buf, off_t offset, size_t count);
 #endif /*DISK_H_*/
